+++ src/main/java/me/despical/classicduels/handlers/PlaceholderManager.java
@@ -19,10 +19,12 @@
 package me.despical.classicduels.handlers;
 
 import me.clip.placeholderapi.expansion.PlaceholderExpansion;
+import me.despical.classicduels.Main;
 import me.despical.classicduels.api.StatsStorage;
 import me.despical.classicduels.arena.Arena;
 import me.despical.classicduels.arena.ArenaRegistry;
 import org.bukkit.entity.Player;
+import org.bukkit.plugin.java.JavaPlugin;
 
 /**
  * @author Despical
@@ -31,6 +33,8 @@ import org.bukkit.entity.Player;
  * Created at 11.10.2020
  */
 public class PlaceholderManager extends PlaceholderExpansion {
+
+	private final Main plugin = JavaPlugin.getPlugin(Main.class);
 
 	@Override
 	public boolean persist() {
@@ -46,7 +50,7 @@ public class PlaceholderManager extends PlaceholderExpansion {
 	}
 
 	public String getVersion() {
-		return "1.0.2";
+		return plugin.getDescription().getVersion();
 	}
 
 	public String onPlaceholderRequest(Player player, String id) {