+++ src/main/java/me/despical/classicduels/handlers/rewards/RewardsFactory.java
@@ -59,7 +59,9 @@ public class RewardsFactory {
 			return;
 		}
 
-		arena.getPlayers().forEach(p -> performReward(p, type));
+		for (Player p : arena.getPlayers()) {
+			performReward(p, type);
+		}
 	}
 
 	public void performReward(Player player, Reward.RewardType type) {
@@ -131,7 +133,10 @@ public class RewardsFactory {
 			}
 		}
 
-		registeredRewards.keySet().forEach(rewardType -> Debugger.debug("[RewardsFactory] Registered {0} {1} rewards!", registeredRewards.get(rewardType), rewardType.name()));
+		for (Reward.RewardType rewardType : registeredRewards.keySet()) {
+			Debugger.debug("[RewardsFactory] Registered {0} {1} rewards!", registeredRewards.get(rewardType), rewardType.name());
+		}
+
 		Debugger.debug("[RewardsFactory] Registered all rewards took {0} ms", System.currentTimeMillis() - start);
 	}
 }